Remind developers to update the docusaurus tarball
This removes the workflow to automatically generate a tarball for
changes to the `docusaurus/` directory in favor of developers being
expected to do this themselves with an automatic reminder in a PR to
inform about it.

This comes from experiencing issues with the automatic compression
workflow, and should be considered a temproary "solution" to an issue
we've faced.

See #44 and #45 for reference.